move grid view into its own file & only scroll when overlfow
